pyproject.toml
[tool.poetry]
name = "biochatter-server"
version = "0.1.0"
description = "the server to host biochatter"
authors = ["fengsh <shaohong.feng.78@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
biochatter = "^0.1.8"
Flask = "^3.0.0"
openai = "0.27.8"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
